update
I have everything working in  "Nikola deploy" and updated the '.blogtxt CSS theme '.  the only issue I am currently running into to is;  for some reason, Nikola is not writing out the logo and favicon to the top level of HTdocs.. or any level for that matter. I mitigated this by placing the images in the correct location in the htdocs folder before upload. The site content on the server will be fine for right now as long as the content on the server isn't deleted or the images removed. I would like help to figure out why Nikola build is not catching these images during the build phase.